deployment/requirements.txt
streamlit
pandas
seaborn
plotly
pillow
numpy
scikit-learn
matplotlib
imblearn
